pending issues resolved
